Tweaked the script.

+++ astro-computer/MPS/cones.resolution.sh
@@ -8,31 +8,45 @@ export OPTIONS="-Dverbose=false"
 #
 # User input
 #
-echo -e "You will be prompted to enter body names, dates, and observed altitudes.\n"
-rm cones.input.txt 2> /dev/null
-KEEP_LOOPING=true
-while [[ "${KEEP_LOOPING}" == "true" ]]; do
-  echo -en "Body name: "
-  read BODY_NAME
-  echo -en "Date (duration format): "
-  read THE_DATE
-  echo -en "Observed Altitude: "
-  read OBS_ALT
-  #
-  echo "Body=${BODY_NAME};Date=${THE_DATE};ObsAlt=${OBS_ALT}" >> cones.input.txt
+PRM_FILE_NAME=
+ASK_USER=true
+if [[ "${1}" != "" ]]; then
+  # File provided on CLI
+  PRM_FILE_NAME=${1}
+  ASK_USER=false
+else
+  PRM_FILE_NAME=cones.input.txt
+fi
 
-  echo -en "More [n]|y ? > "
-  read REPLY
-  if [[ ${REPLY} =~ ^(yes|y|Y)$ ]]; then
-    echo -e "Ok, another one."
-  else
-    echo -e "Now proceeding."
-    KEEP_LOOPING=false
-  fi
-done
+if [[ "${ASK_USER}" == "true" ]]; then
+  echo -e "You will be prompted to enter body names, dates, and observed altitudes.\n"
+  rm cones.input.txt 2> /dev/null
+  KEEP_LOOPING=true
+  while [[ "${KEEP_LOOPING}" == "true" ]]; do
+    echo -en "Body name: "
+    read BODY_NAME
+    echo -en "Date (duration format): "
+    read THE_DATE
+    echo -en "Observed Altitude: "
+    read OBS_ALT
+    #
+    echo "Body=${BODY_NAME};Date=${THE_DATE};ObsAlt=${OBS_ALT}" >> cones.input.txt
+
+    echo -en "More [n]|y ? > "
+    read REPLY
+    if [[ ${REPLY} =~ ^(yes|y|Y)$ ]]; then
+      echo -e "Ok, another one."
+    else
+      echo -e "Now proceeding."
+      KEEP_LOOPING=false
+    fi
+  done
+else
+  echo -e "Using data from ${PRM_FILE_NAME} (if found)"
+fi
 #
 # Resolution
 #
-java -classpath ${CP} ${OPTIONS} mps.pg.PlayGround09 cones.input.txt
+java -classpath ${CP} ${OPTIONS} mps.pg.PlayGround09 ${PRM_FILE_NAME}
 #
 echo -e "Done"
